added right size icon
