+++ src/hooks/usePools.ts
@@ -80,61 +80,5 @@ export function usePools(wallet) {
     }
   }, []);
 
-  // useEffect(() => {
-  //   async function fetchPools() {
-  //     console.log("in fetch pools");
-  //     let { perpetual_program } = await getPerpetualProgramAndProvider(wallet);
-
-  //     let fetchedPools = await perpetual_program.account.pool.all();
-
-  //     let poolInfos: Pool[] = [];
-
-  //     Object.values(fetchedPools).forEach(async (pool) => {
-  //       let custodyAccounts = [];
-  //       Object.values(pool.account.tokens).forEach((token) => {
-  //         custodyAccounts.push(token.custody.toString());
-  //       });
-
-  //       let fetchedCustodies =
-  //         await perpetual_program.account.custody.fetchMultiple(
-  //           custodyAccounts
-  //         );
-  //       console.log("custody example", fetchedCustodies[0]);
-
-  //       // setCustodies({
-  //       //   ...custodies,
-  //       //   [pool.publicKey.toString()]: fetchedCusto,
-  //       // });
-  //       let custodyInfos: TokenCustody[] = [];
-
-  //       Object.values(fetchedCustodies).forEach((custody) => {
-  //         custodyInfos.push({
-  //           mint: custody.mint.toString(),
-  //         });
-  //       });
-
-  //       let poolObject = {
-  //         poolName: pool.account.name,
-  //         // tokens: custodyInfos,
-  //       };
-  //       console.log("got pool obj", poolObject);
-  //       poolInfos.push(poolObject);
-  //     });
-
-  //     let poolObject = {
-  //       poolName: "bro",
-  //     };
-  //     poolInfos.push(poolObject);
-
-  //     setPools(poolInfos);
-  //   }
-  //   console.log("running effect", pools);
-  //   if (pools.length === 0) {
-  //     console.log("about to fetch");
-  //     fetchPools();
-  //     console.log("done fetching", pools);
-  //   }
-  // }, []);
-
   return { pools };
 }
